Small clean up
  * Move input-processor-related APIs to `deprecated` file
  * Move building of type-maps to Schema class
