+++ lib/dry/validation/schema.rb
@@ -12,7 +12,7 @@ require 'dry/validation/messages'
 require 'dry/validation/error_compiler'
 require 'dry/validation/hint_compiler'
 
-require 'dry/validation/input_processor_compiler'
+require 'dry/validation/schema/deprecated'
 
 module Dry
   module Validation
@@ -33,7 +33,6 @@ module Dry
       setting :type_map, {}
 
       setting :input_processor, :noop
-
       setting :input_processor_map, {
         sanitizer: InputProcessorCompiler::Sanitizer.new,
         json: InputProcessorCompiler::JSON.new,
@@ -56,6 +55,37 @@ module Dry
 
       def self.set_registry!
         config.registry = PredicateRegistry[self, config.predicates]
+      end
+
+      def self.registry
+        config.registry
+      end
+
+      def self.build_type_map(type_specs, category = config.input_processor)
+        type_specs.each_with_object({}) do |(name, spec), result|
+          result[name] =
+            case spec
+            when Hash
+              Types["#{category}.hash"].symbolized(spec)
+            when Array
+              if spec.size == 1 && spec[0].is_a?(Hash)
+                member = Types["#{category}.hash"].symbolized(build_type_map(spec[0], category))
+                Types["#{category}.array"].member(member)
+              else
+                spec
+                  .map { |id| id.is_a?(Symbol) ? Types["#{category}.#{id}"] : id }
+                  .reduce(:|)
+              end
+            when Symbol
+              Types["#{category}.#{spec}"]
+            else
+              spec
+            end
+        end
+      end
+
+      def self.type_map
+        config.type_map
       end
 
       def self.predicates(predicate_set = nil)
@@ -100,10 +130,6 @@ module Dry
         [:schema, self]
       end
 
-      def self.type_map
-        config.type_map
-      end
-
       def self.rules
         config.rules
       end
@@ -143,33 +169,8 @@ module Dry
         @hint_compiler ||= HintCompiler.new(messages, rules: rule_ast)
       end
 
-      def self.input_processor
-        @input_processor ||=
-          begin
-            if type_map.size > 0
-              Types["#{config.input_processor}.hash"].symbolized(type_map)
-            elsif input_processor_compiler
-              input_processor_compiler.(rule_ast)
-            else
-              NOOP_INPUT_PROCESSOR
-            end
-          end
-      end
-
-      def self.input_processor_ast(type)
-        config.input_processor_map.fetch(type).schema_ast(rule_ast)
-      end
-
-      def self.input_processor_compiler
-        @input_processor_comp ||= config.input_processor_map[config.input_processor]
-      end
-
       def self.rule_ast
         @rule_ast ||= config.rules.flat_map(&:rules).map(&:to_ast)
-      end
-
-      def self.registry
-        config.registry
       end
 
       def self.default_options
